+++ packages/axios-extension/src/abap/lrep-service.ts
@@ -60,6 +60,7 @@ export interface MergedAppDescriptor {
                 final: boolean;
             };
         }[];
+        requests?: unknown[];
     };
 }