+++ packages/jest-file-matchers/src/index.ts
@@ -1,1 +1,2 @@
 export * from './matchers';
+export * from './matchers/types';