+++ packages/preview-middleware-client/test/__mock__/sap/ui/core/Control.ts
@@ -1,0 +1,3 @@
+export default class Control {
+    getManifest = jest.fn();
+}