+++ packages/preview-middleware-client/test/unit/adp/init-dialogs.test.ts
@@ -1,8 +1,16 @@
 import type UI5Element from 'sap/ui/core/Element';
-import { DialogNames, handler, initDialogs } from '../../../src/adp/init-dialogs';
-import XMLView from 'mock/sap/ui/core/mvc/XMLView';
+import { RTAOptions } from 'sap/ui/rta/RuntimeAuthoring';
+import type RuntimeAuthoring from 'sap/ui/rta/RuntimeAuthoring';
+
+import { sapMock } from 'mock/window';
+import Fragment from 'mock/sap/ui/core/Fragment';
 import Controller from 'mock/sap/ui/core/mvc/Controller';
-import rtaMock from 'mock/sap/ui/rta/RuntimeAuthoring';
+import RuntimeAuthoringMock from 'mock/sap/ui/rta/RuntimeAuthoring';
+
+import { DialogNames, handler, initDialogs } from '../../../src/adp/init-dialogs';
+import AddFragment from '../../../src/adp/controllers/AddFragment.controller';
+import ControllerExtension from '../../../src/adp/controllers/ControllerExtension.controller';
+import ExtensionPoint from '../../../src/adp/controllers/ExtensionPoint.controller';
 
 describe('Dialogs', () => {
     describe('initDialogs', () => {
@@ -12,23 +20,42 @@ describe('Dialogs', () => {
 
         test('adds a new item to the context menu', () => {
             const addMenuItemSpy = jest.fn();
+            const rtaMock = new RuntimeAuthoringMock({} as RTAOptions);
             rtaMock.getDefaultPlugins.mockReturnValueOnce({
                 contextMenu: {
                     addMenuItem: addMenuItemSpy
                 }
             });
-            initDialogs(rtaMock);
+            initDialogs(rtaMock as unknown as RuntimeAuthoring);
             expect(addMenuItemSpy).toHaveBeenCalledTimes(2);
         });
 
         test('addMenuItem handler function', async () => {
-            Controller.create.mockResolvedValue({ overlays: {}, rta: {} });
+            sapMock.ui.version = '1.120.4';
+            Controller.create.mockResolvedValue({ overlays: {}, rta: { 'yes': 'no' } });
+            const rtaMock = new RuntimeAuthoringMock({} as RTAOptions);
 
-            await handler({} as unknown as UI5Element, rtaMock, DialogNames.ADD_FRAGMENT);
-            await handler({} as unknown as UI5Element, rtaMock, DialogNames.CONTROLLER_EXTENSION);
-            await handler({} as unknown as UI5Element, rtaMock, DialogNames.ADD_FRAGMENT_AT_EXTENSION_POINT);
+            AddFragment.prototype.setup = jest.fn();
+            ControllerExtension.prototype.setup = jest.fn();
+            ExtensionPoint.prototype.setup = jest.fn();
 
-            expect(XMLView.create).toHaveBeenCalledTimes(3);
+            await handler(
+                {} as unknown as UI5Element,
+                rtaMock as unknown as RuntimeAuthoring,
+                DialogNames.ADD_FRAGMENT
+            );
+            await handler(
+                {} as unknown as UI5Element,
+                rtaMock as unknown as RuntimeAuthoring,
+                DialogNames.CONTROLLER_EXTENSION
+            );
+            await handler(
+                {} as unknown as UI5Element,
+                rtaMock as unknown as RuntimeAuthoring,
+                DialogNames.ADD_FRAGMENT_AT_EXTENSION_POINT
+            );
+
+            expect(Fragment.load).toHaveBeenCalledTimes(3);
         });
     });
 });