+++ packages/preview-middleware-client/test/unit/flp/init.test.ts
@@ -2,7 +2,7 @@ import { init, registerComponentDependencyPaths, registerSAPFonts, setI18nTitle 
 import IconPoolMock from 'mock/sap/ui/core/IconPool';
 import { mockBundle } from 'mock/sap/base/i18n/ResourceBundle';
 import { fetchMock, sapMock } from 'mock/window';
-import type { RTAPlugin, StartAdaptation } from 'sap/ui/rta/api/startAdaptation';
+import type { InitRtaScript, RTAPlugin, StartAdaptation } from 'sap/ui/rta/api/startAdaptation';
 
 describe('flp/init', () => {
     test('registerSAPFonts', () => {
@@ -78,6 +78,7 @@ describe('flp/init', () => {
     describe('init', () => {
         beforeEach(() => {
             sapMock.ushell.Container.attachRendererCreatedEvent.mockReset();
+            sapMock.ui.require.mockReset();
         });
 
         test('nothing configured', async () => {
@@ -91,6 +92,7 @@ describe('flp/init', () => {
                 layer: 'CUSTOMER_BASE',
                 pluginScript: 'my/script'
             };
+            sapMock.ui.version = '1.84.50';
             await init({ flex: JSON.stringify(flexSettings) });
             expect(sapMock.ushell.Container.attachRendererCreatedEvent).toBeCalled();
             expect(sapMock.ushell.Container.createRenderer).toBeCalledWith(undefined, true);
@@ -121,6 +123,44 @@ describe('flp/init', () => {
             const plugnScriptMock = jest.fn();
             requireCb(startAdpMock, plugnScriptMock);
             expect(startAdpMock).toBeCalledWith(expect.anything(), plugnScriptMock);
+        });
+
+        test('flex configured & ui5 version is 1.71.60', async () => {
+            const flexSettings = {
+                layer: 'CUSTOMER_BASE',
+                pluginScript: 'my/script'
+            };
+            sapMock.ui.version = '1.71.60';
+            await init({ flex: JSON.stringify(flexSettings) });
+            expect(sapMock.ushell.Container.attachRendererCreatedEvent).toBeCalled();
+            expect(sapMock.ushell.Container.createRenderer).toBeCalledWith(undefined, true);
+            const rendererCb = sapMock.ushell.Container.attachRendererCreatedEvent.mock
+                .calls[0][0] as () => Promise<void>;
+
+            // testing the nested callbacks
+            const mockService = {
+                attachAppLoaded: jest.fn()
+            };
+            sapMock.ushell.Container.getServiceAsync.mockResolvedValueOnce(mockService);
+
+            await rendererCb();
+            expect(mockService.attachAppLoaded).toBeCalled();
+            const loadedCb = mockService.attachAppLoaded.mock.calls[0][0] as (event: unknown) => void;
+
+            loadedCb({ getParameter: () => {} });
+            expect(sapMock.ui.require).toBeCalledWith(
+                ['open/ux/preview/client/flp/initRta', flexSettings.pluginScript],
+                expect.anything()
+            );
+
+            const requireCb = sapMock.ui.require.mock.calls[0][1] as (
+                initRta: InitRtaScript,
+                pluginScript?: RTAPlugin
+            ) => Promise<void>;
+            const initRtaMock = jest.fn();
+            const plugnScriptMock = jest.fn();
+            await requireCb(initRtaMock, plugnScriptMock);
+            expect(initRtaMock).toBeCalledWith(expect.anything(), plugnScriptMock);
         });
     });
 });