+++ packages/preview-middleware-client/types/sap.ui.rta.d.ts
@@ -44,6 +44,12 @@ declare module 'sap/ui/rta/command/FlexCommand' {
     import type Change from 'sap/ui/fl/Change';
 
     interface FlexCommand extends Omit<BaseCommand, 'getCommands'> {
+        _oPreparedChange?: {
+            _oDefinition: {
+                moduleName: string;
+            };
+            setModuleName(moduleName: string): void;
+        };
         getPreparedChange(): Change;
         getCommands(): FlexCommand[];
     }
@@ -109,6 +115,18 @@ declare module 'sap/ui/rta/command/OutlineService' {
     export default OutlineService;
 }
 
+declare module 'sap/ui/fl/FakeLrepConnector' {
+    export default class FakeLrepConnector {
+        static enableFakeConnector: () => void;
+    }
+}
+
+declare module 'sap/ui/fl/LrepConnector' {
+    export default class LrepConnector {
+        loadChanges(): Promise<any>;
+    }
+}
+
 declare module 'sap/ui/rta/RuntimeAuthoring' {
     import type Event from 'sap/ui/base/Event';
     import type Component from 'sap/ui/core/Component';
@@ -117,12 +135,17 @@ declare module 'sap/ui/rta/RuntimeAuthoring' {
     import type ContextMenu from 'sap/ui/dt/plugin/ContextMenu';
     import type { Layer } from 'sap/ui/fl';
     import type { Scenario } from 'sap/ui/fl/Scenario';
+    import type Control from 'sap/ui/core/Control';
 
     type Manifest = {
         [key: string]: unknown;
         'sap.app': {
             [key: string]: string;
             id: string;
+        };
+        'sap.ui5': {
+            [key: string]: string;
+            flexEnabled?: boolean;
         };
     };
 
@@ -162,7 +185,19 @@ declare module 'sap/ui/rta/RuntimeAuthoring' {
         generator: string;
     }
 
-    interface RuntimeAuthoring {
+    export interface RTAOptions {
+        [key: string]: any;
+        flexSettings: FlexSettings;
+        rootControl: Control;
+        validateAppVersion: boolean;
+    }
+
+    export default class RuntimeAuthoring {
+        constructor(_: RTAOptions) {}
+
+        destroy: () => void;
+        start: () => Promise<void>;
+        attachEvent: (name: string, fn: () => any) => void;
         attachSelectionChange(handler: (event: SelectionChangeEvent) => void): void;
         attachModeChanged: (handler: (event: Event) => void) => void;
         attachUndoRedoStackModified: (handler: (event: Event) => Promise<void>) => void;
@@ -176,15 +211,14 @@ declare module 'sap/ui/rta/RuntimeAuthoring' {
             getManifest(): Manifest;
         } & Component;
     }
-
-    export default RuntimeAuthoring;
 }
 
 declare module 'sap/ui/rta/api/startAdaptation' {
     import type RuntimeAuthoring from 'sap/ui/rta/RuntimeAuthoring';
 
-    export type RTAPlugin = (rta: RuntimeAuthoring) => void;
+    export type RTAPlugin = (rta: RuntimeAuthoring) => Promise<void> | void;
     export type StartAdaptation = (options: object, plugin?: RTAPlugin) => void;
+    export type InitRtaScript = (options: RTAOptions, pluginScript: RTAPlugin) => Promise<void>;
 
     const startAdaptation: StartAdaptation;