+++ packages/ui5-info/src/index.ts
@@ -1,2 +1,5 @@
-export { getUI5Versions } from './ui5-info';
-export type { UI5VersionFilterOptions, UI5Version } from './types';
+export { defaultVersion, minUi5VersionSupportingCodeAssist, latestVersionString } from './constants';
+export type { UI5Theme, UI5Version, UI5VersionFilterOptions } from './types';
+export { getUI5Versions } from './ui5-version-info';
+export { supportedUi5VersionFallbacks } from './ui5-version-fallback';
+export { ui5ThemeIds, getDefaultUI5Theme, getUi5Themes } from './ui5-theme-info';