+++ examples/simple-generator/CHANGELOG.md
@@ -1,5 +1,77 @@
 # @sap-ux/generator-simple-fe
 
+## 0.3.29
+
+### Patch Changes
+
+-   @sap-ux/system-access@0.3.17
+
+## 0.3.28
+
+### Patch Changes
+
+-   Updated dependencies [7964a7ba]
+    -   @sap-ux/fiori-elements-writer@0.18.8
+
+## 0.3.27
+
+### Patch Changes
+
+-   @sap-ux/fiori-elements-writer@0.18.7
+
+## 0.3.26
+
+### Patch Changes
+
+-   Updated dependencies [811c4324]
+    -   @sap-ux/btp-utils@0.14.2
+    -   @sap-ux/axios-extension@1.11.3
+    -   @sap-ux/system-access@0.3.16
+
+## 0.3.25
+
+### Patch Changes
+
+-   c15435b6: fix: remove engines pnpm from package.json
+-   Updated dependencies [c15435b6]
+    -   @sap-ux/fiori-freestyle-writer@0.21.4
+    -   @sap-ux/fiori-elements-writer@0.18.6
+    -   @sap-ux/axios-extension@1.11.2
+    -   @sap-ux/system-access@0.3.15
+    -   @sap-ux/btp-utils@0.14.1
+    -   @sap-ux/logger@0.5.1
+
+## 0.3.24
+
+### Patch Changes
+
+-   Updated dependencies [efd2f6d4]
+    -   @sap-ux/axios-extension@1.11.1
+    -   @sap-ux/system-access@0.3.14
+
+## 0.3.23
+
+### Patch Changes
+
+-   Updated dependencies [0f6e0e1b]
+    -   @sap-ux/axios-extension@1.11.0
+    -   @sap-ux/btp-utils@0.14.0
+    -   @sap-ux/system-access@0.3.13
+
+## 0.3.22
+
+### Patch Changes
+
+-   @sap-ux/fiori-elements-writer@0.18.5
+
+## 0.3.21
+
+### Patch Changes
+
+-   Updated dependencies [64f9c513]
+    -   @sap-ux/axios-extension@1.10.2
+    -   @sap-ux/system-access@0.3.12
+
 ## 0.3.20
 
 ### Patch Changes