+++ packages/logger/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sap-ux/logger
 
+## 0.5.1
+
+### Patch Changes
+
+-   c15435b6: fix: remove engines pnpm from package.json
+
 ## 0.5.0
 
 ### Minor Changes